tf1_requirements.txt
absl-py==1.3.0
astor==0.8.1
cycler==0.11.0
fonttools==4.38.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.51.1
h5py==2.10.0
imageio==2.22.4
importlib-metadata==5.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
opt-einsum==3.3.0
packaging==21.3
Pillow==9.3.0
protobuf==3.20.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==2.1.1
tifffile==2021.11.2
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
